RRD-15: Update Navigation folders structure
